+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Version 1.1.0 (?)
 
 * [new] Add Bean Validation support to Morphia (at pre-persist)
+* [new] Add `exists()` and `count()` in `BaseMorphiaRepository`
 * [chg] Update for SeedStack 16.4
+* [brk] remove `do*()` methods in `BaseMorphiaRepository`
 
 # Version 1.0.1 (2016-02-09)